Ensure array index is integer

+++ includes/civicrm/cwps-civicrm-custom-field.php
@@ -266,7 +266,7 @@ class CiviCRM_WP_Profile_Sync_CiviCRM_Custom_Field {
 		foreach ( $result['values'] as $item ) {
 			foreach ( $item as $key => $value ) {
 				if ( substr( $key, 0, 7 ) == 'custom_' ) {
-					$index = str_replace( 'custom_', '', $key );
+					$index = (int) str_replace( 'custom_', '', $key );
 					$contact_data[ $index ] = $value;
 				}
 			}
